requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
jinja2==3.1.3
starlette==0.36.1
itsdangerous==2.1.2
python-jose==3.3.0
email-validator==2.1.0.post1 